Default opacity 1.0 via Cesium3dTilesStyleMixin - make mixin observable
