+++ lib/ModelMixins/Cesium3dTilesStyleMixin.ts
@@ -1,11 +1,32 @@
 import Cesium3DTileStyle from "terriajs-cesium/Source/Scene/Cesium3DTileStyle";
 import AbstractConstructor from "../Core/AbstractConstructor";
 import isDefined from "../Core/isDefined";
-import Model from "../Models/Definition/Model";
+import Model, { BaseModel } from "../Models/Definition/Model";
 import Cesium3dTilesTraits from "../Traits/TraitsClasses/Cesium3dTilesTraits";
 import clone from "terriajs-cesium/Source/Core/clone";
-import { computed, override, toJS } from "mobx";
+import { computed, makeObservable, override, runInAction, toJS } from "mobx";
 import Color from "terriajs-cesium/Source/Core/Color";
+import LoadableStratum from "../Models/Definition/LoadableStratum";
+import StratumOrder from "../Models/Definition/StratumOrder";
+
+class Cesium3dTilesStyleStratum extends LoadableStratum(Cesium3dTilesTraits) {
+  constructor(...args: any[]) {
+    super(...args);
+    makeObservable(this);
+  }
+
+  duplicateLoadableStratum(model: BaseModel): this {
+    return new Cesium3dTilesStyleStratum(model) as this;
+  }
+
+  @computed
+  get opacity() {
+    return 1.0;
+  }
+}
+
+// Register the I3SStratum
+StratumOrder.instance.addLoadStratum(Cesium3dTilesStyleStratum.name);
 
 type BaseType = Model<Cesium3dTilesTraits>;
 
@@ -17,6 +38,13 @@ function Cesium3dTilesStyleMixin<T extends AbstractConstructor<BaseType>>(
   abstract class Cesium3dTilesStyleMixin extends Base {
     constructor(...args: any[]) {
       super(...args);
+      makeObservable(this);
+      runInAction(() => {
+        this.strata.set(
+          Cesium3dTilesStyleStratum.name,
+          new Cesium3dTilesStyleStratum()
+        );
+      });
     }
 
     get hasCesium3dTilesStyleMixin() {
